% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_choose_mode.R
\name{choose_mode-methods}
\alias{choose_mode-methods}
\alias{print.choose_mode}
\alias{plot.choose_mode}
\title{Methods for 'choose_mode' objects}
\usage{
\method{print}{choose_mode}(x, ...)

\method{plot}{choose_mode}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"choose_mode"}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Methods for 'choose_mode' objects
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@live.com}>
}
