% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_choose_mode.R
\name{choose_mode}
\alias{choose_mode}
\title{Choose the Mode of the Modified SDL Regression}
\usage{
choose_mode(
  object,
  grid = -5:5,
  trace = TRUE,
  plot = TRUE,
  control = sdl_control(...),
  ...
)
}
\arguments{
\item{object}{an \code{"sdlrm"}object.}

\item{grid}{grid of values that will be used to evaluate the profile log-likelihood function.}

\item{trace}{logical; if \code{TRUE}, a summary with the profile log-likelihood value, the AIC,
and the BIC of the fit is displayed.}

\item{plot}{logical; if \code{TRUE}, a graph of the profiled log-likelihood evaluated in the
considered grid of values is shown.}

\item{control}{a list of control arguments specified via \code{\link{sdl_control}}.}

\item{...}{further arguments passed to \code{\link{sdl_control}}.}
}
\value{
An object of class \code{"choose_mode"}. More specifically, it returns a list in which
    each element consists of the fit of the modified SDL regression with each value of the mode
    specified in \code{grid}.
}
\description{
Estimation of the mode of the modified SDL distribution in the fit of the modified SDL regression
    model via profile log-likelihood.
}
\examples{

# Data set: pss (for description run ?pss)
barplot(table(pss$difference), xlab = "PSS index difference", ylab = "Frequency")
boxplot(pss$difference ~ pss$group, xlab = "Group", ylab = "PSS index difference")

# Fit with a model only for the mean with xi = 0 (default)
fit0 <- sdlrm(difference ~ group, data = pss)

# Choosing the mode on the grid {-10, -9, ..., 0, ..., 9, 10}
fit <- choose_mode(fit0, grid = -10:10)

# Class
class(fit)

# It is possible to recovery the plot:
plot(fit)

# and the trace:
fit

# Fit with xi = 1
fit[[1]]

}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
