% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdl_diagnostic.R
\name{envel_sdl}
\alias{envel_sdl}
\title{Envelope Graph of the Sdl Regression Model}
\usage{
envel_sdl(object, R = 99, control = sdl_control())
}
\arguments{
\item{object}{object of class 'sdlrm'.}

\item{R}{number of replicates.}

\item{control}{a list of control arguments specified via \code{sdl_control}.}
}
\value{
Returns the normal envelope simulated probability plot of
Pearson residuals and random quantile residuals, respectively.
}
\description{
Provides the normal probability plot with simulated
envelope of Pearson residuals and randomized quantile residuals
resulting from the Sdl regression model fit.
}
\references{
Medeiros, R. & Bourguignon, M. (2020).
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
