% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_sdl_methods.R
\name{plot.sdlrm}
\alias{plot.sdlrm}
\title{Diagnostic Plots for the Modified SDL Regression}
\usage{
\method{plot}{sdlrm}(
  x,
  which = 1:2,
  type = c("quantile", "pearson", "response"),
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  ...
)
}
\arguments{
\item{x}{an object of class \code{"sdlrm"}.}

\item{which}{numeric; if a subset of the plots is required, specify a subset
of the numbers \code{1:6}.}

\item{type}{character; specifies which residual should be produced in the
envelope plot. The available options are \code{"quantile"} (default),
\code{"pearson"}, and \code{"response"} (raw residuals, y - mu).}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Six plots (selectable by \code{which}) are currently available:
a plot of residuals against fitted values, a plot of residuals against
the observation indices, a Normal Q-Q plot, a barplot with comparisons of the
observed and fitted frequencies, a plot of the sample autocorelations
of the residuals, and a plot of the sample partial autocorelations
of the residuals.
}
\examples{
## Data set: pss (for description run ?pss)
barplot(table(pss$difference), xlab = "PSS index difference", ylab = "Frequency")
boxplot(pss$difference ~ pss$group, xlab = "Group", ylab = "PSS index difference")

## Fit with a model only for the mean (mode = 1)
fit <- sdlrm(difference ~ group, data = pss, xi = 1)

## Available plots (using the randomized quantile residuals):
# Residuals versus fitted values
plot(fit, which = 1)

# Residuals versus observation indices
plot(fit, which = 2)

# Normal Q-Q plot
plot(fit, which = 3)

# Observed versus fitted frequencies
plot(fit, which = 4)

# Sample autocorelation function of residuals
plot(fit, which = 5)

# Sample partial autocorelation of residuals
plot(fit, which = 6)
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
