% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdl_diagnostic.R
\name{rqr_sdl}
\alias{rqr_sdl}
\title{Randomized quantile residuals in the Sdl regression model}
\usage{
rqr_sdl(y, mu, phi)
}
\arguments{
\item{y}{vector; observed values of the response variable.}

\item{mu}{fitted values.}

\item{phi}{fitted values for the dispersion parameter.}
}
\value{
Randomized quantile residuals resulting from the fit of
the Sdl regression model.
}
\description{
Randomized quantile residuals resulting from the fit of
the Sdl regression model.
}
\references{
Medeiros, R. & Bourguignon, M. (2020). \cr
Dunn, P. K., & Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236-244.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
