% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdl_diagnostic.R
\name{expected_sdl}
\alias{expected_sdl}
\alias{expect_sdl}
\title{Expected Frequencies by the Sdl Regression Model}
\usage{
expect_sdl(y, mu, phi)
}
\arguments{
\item{y}{vector of the observed values of the response variable.}

\item{mu}{vector of the fitted values.}

\item{phi}{vector of the fitted values for the dispersion parameter.}
}
\value{
Expected frequencies for a fit of the Sdl regression model
to a dataset.
}
\description{
Provides the expected frequencies for a fit of
the Sdl regression model according to the definition given in
Kleiber and Zeileis (2016).
}
\examples{

# Sample size
n <- 100
# Covariates
X <- cbind(rep(1,n), runif(n, 0,1), runif(n,0,1))
Z <- cbind(rep(1,n), runif(n, 0,1), runif(n,0,1))

# Parameters and relations
beta <- c(1, 1.2, 0.2)
gama <- c(2, 1.5, 1.2)

mu <- exp(X\%*\%beta)
phi <- exp(Z\%*\%gama)

y <- rsdl(n, mu, phi)

expect_sdl(y, mu, phi)

}
\references{
Medeiros, R. & Bourguignon, M. (2020).  \cr
Kleiber, C., & Zeileis, A. (2016). Visualizing count data regressions using rootograms. The American Statistician, 70, 296-303
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
