% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdl_distribution.R
\name{sdl}
\alias{sdl}
\alias{dsdl}
\alias{psdl}
\alias{qsdl}
\alias{rsdl}
\title{The Skew Discrete Laplace (Sdl) Distribution}
\usage{
dsdl(x, mu, phi)

psdl(q, mu, phi, lower.tail = TRUE)

qsdl(p, mu, phi, lower.tail = TRUE)

rsdl(n, mu, phi)
}
\arguments{
\item{x}{vector of integer quantiles.}

\item{mu}{vector of means.}

\item{phi}{vector of dispersion indexes (greather than \code{abs(mu)}).}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P(X <= x)}, otherwise, \code{P(X > x)}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\value{
\code{dsdl} returns the probability function, \code{psdl}
gives the distribution function, \code{qsdl} gives the quantile function,
and \code{rsdl} generates random observations.
}
\description{
Probability mass function, distribution function,
quantile function and random generation for the Sdl distribution
with mean \code{mu} and dispersion parameter \code{phi}.
}
\details{
The Sdl distribution was proposed by Kozubowski and Inusah (2006). This
probability distribution has some attractive features as the support
in integers \code{Z = {..., -2, -1, 0, 1, 2, ...}}; mathematical expressions in closed-form
for the cumulative distribution function, quantile function, characteristic function,
ordinal and absolute moments; assumes a skewed-left and skewed-right distribution,
and also symmetry; in addition, it can be represented stochastically as the
difference between two independent random variables, both with geometric distribution.

This set of functions represents the probability function, the cumulative distribution
function, quantile function and a random number generator for the Sdl distribution
parameterized in terms of the mean and a dispersion parameter. This new
parameterization was proposed by Medeiros and Bourguignin (2020).
}
\examples{

### Probability function ###

# Parameters
mu <- c(-2, 0, 2); phi <- 3
# Skewed-left distribution (mu < 0)
plot(-15:15, dsdl(-15:15, mu[1], phi),
      type = "h", xlab = "x", ylab = "Pmf")

# Simmetric distribution (mu = 0)
plot(-15:15, dsdl(-15:15, mu[2], phi),
      type = "h", xlab = "x", ylab = "Pmf")

# Skewed-right distribution (mu > 0)
plot(-15:15, dsdl(-15:15, mu[3], phi),
      type = "h", xlab = "x", ylab = "Pmf")

### Difference between paired samples of non-negative observations  ###

# Parameters
mu <- 2; phi <- 4

# Paired samples of a pre-post treatment experimental study
before <- rgeom(1000, 2 / (2 + phi - mu))
after <- rgeom(1000, 2 / (2 + phi + mu))

# Response variable
y <- after - before

# Barplot
obj <- barplot(prop.table(table(y)),
               xlab = "Response",
               ylab = "Proportion",
               col = "white",
               ylim = c(0, dsdl(0, mu, phi)))

# Sdl model for the differences
points(obj, dsdl(sort(unique(y)), mu, phi), col = "red", pch = 16)

}
\references{
Medeiros, R. and Bourguignon, M. (2020).

Kozubowski, T. J., and Inusah, S. (2006). A skew Laplace distribution on integers. Annals of the Institute of Statistical Mathematics, 58, 555--571.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
