% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdl_methods.R
\name{sdlrm-methods}
\alias{sdlrm-methods}
\alias{plot.sdlrm}
\alias{logLik.sdlrm}
\alias{AIC.sdlrm}
\alias{coef.sdlrm}
\alias{vcov.sdlrm}
\alias{model.matrix.sdlrm}
\title{Methods for "sdlrm" objects}
\usage{
\method{plot}{sdlrm}(x, ...)

\method{logLik}{sdlrm}(object)

\method{AIC}{sdlrm}(object, ..., k = 2)

\method{coef}{sdlrm}(object, what = c("all", "mean", "dispersion"))

\method{vcov}{sdlrm}(object)

\method{model.matrix}{sdlrm}(object, matrix = c("all", "mean", "dispersion"))
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{object, x}{an object of class \code{"sdlrm"}.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}

\item{what}{a character indicating which parameter coefficients are
required, parameters for the \code{"mean"} or for the
\code{"dispersion"} model. If \code{"all"} (default), a list with
coefficients for the \code{mean} and for the \code{dispersion}
model is returned.}

\item{matrix}{a character indicating which model matrix is
required, the model matrix for the mean (\code{"mean"}) or for the
dispersion parameter (\code{"dispersion"}). If \code{"all"} (default), a list with
with both matrices are returned.}
}
\description{
Methods for "sdlrm" objects
}
\references{
Medeiros, R. M. R., Bourguignon, M. (2020). Mean and
    dispersion regression model for paired count data with
    application to quality of life in prisons.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
