% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_sdl_methods.R
\name{sdlrm-methods}
\alias{sdlrm-methods}
\alias{model.frame.sdlrm}
\alias{model.matrix.sdlrm}
\alias{coef.sdlrm}
\alias{vcov.sdlrm}
\alias{logLik.sdlrm}
\alias{AIC.sdlrm}
\alias{print.sdlrm}
\alias{summary.sdlrm}
\alias{print.summary.sdlrm}
\title{Methods for \code{"sdlrm"} objects.}
\usage{
\method{model.frame}{sdlrm}(formula, ...)

\method{model.matrix}{sdlrm}(object, parm = c("mean", "dispersion"), ...)

\method{coef}{sdlrm}(object, parm = c("mean", "dispersion", "full"), ...)

\method{vcov}{sdlrm}(object, parm = c("mean", "dispersion", "full"), ...)

\method{logLik}{sdlrm}(object, ...)

\method{AIC}{sdlrm}(object, ..., k = 2)

\method{print}{sdlrm}(x, digits = getOption("digits"), ...)

\method{summary}{sdlrm}(object, ...)

\method{print}{summary.sdlrm}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{formula}{a model \link{formula} or \link{terms} object or an \code{"sdlrm"} object.}

\item{...}{further arguments passed to or from other methods.}

\item{parm}{a character indicating which group of parameters is to be considered in the function.
The options are \code{"mean"} (default) to obtain the output in relation to the coefficients
associated with the mean, \code{"dispersion"} to obtain the output in relation to
the coefficients associated with the dispersion, or \code{"full"} (when applicable)
to consider all the parameters,}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC.}

\item{x, object}{an object of class \code{"sdlrm"}.}

\item{digits}{a non-null value for digits specifies the minimum number of significant digits to
be printed in values. The default, \code{getOption("digits")}.}
}
\description{
Methods for \code{"sdlrm"} objects.
}
\examples{
# Data set: pss (for description run ?pss)
barplot(table(pss$difference), xlab = "PSS index difference", ylab = "Frequency")
boxplot(pss$difference ~ pss$group, xlab = "Group", ylab = "PSS index difference")

# Fit a double model (mode = 1)
fit <- sdlrm(difference ~ group | group, data = pss, xi = 1)

# Print
fit

# Summary
summary(fit)

# Coef
coef(fit)
coef(fit, parm = "dispersion")
coef(fit, parm = "full")

# vcov
vcov(fit)
vcov(fit, parm = "dispersion")
vcov(fit, parm = "full")

# Log-likelihood value
logLik(fit)

# AIC and BIC
AIC(fit)
AIC(fit, k = log(fit$nobs))

# Model matrices
model.matrix(fit)
model.matrix(fit, "dispersion")
}
\author{
Rodrigo M. R. de Medeiros <\email{rodrigo.matheus@ufrn.br}>
}
