% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdl_regression.R
\name{sdlrm}
\alias{sdlrm}
\alias{fit_sdl}
\title{Sdl Regression for Integer-Valued Data}
\usage{
sdlrm(
  formula,
  data,
  link = NULL,
  disp_test = FALSE,
  control = sdl_control(...),
  ...
)

fit_sdl(y, X, Z = NULL, link = "log", control = sdl_control())
}
\arguments{
\item{formula}{description of the model, of type \code{y ~ x} for covariates in mean only, or \code{y ~ x | z} to enter covariates in the dispersion parameter.}

\item{data}{an optional data frame containing the variables in the formula. By default the variables are taken from environment(formula).}

\item{link}{character specification of the link function for the dispersion parameter. The links \code{"log"} (default) \code{"sqrt"} and \code{"identity"} can be used.}

\item{disp_test}{logical, if TRUE, the function \code{sdlrm} returns the test for constant dispersion.}

\item{control}{a list of control arguments specified via \code{sdl_control}.}

\item{...}{arguments passed to \code{sdl_control}.}

\item{y}{a integer-valued vector of the response variable. This argument is used in the \code{fit_sdl} function to get the estimates and required quantities that will return in the object resulting from the sdlrm function.}

\item{X, Z}{model matrices associated with the mean and the dispersion parameter, respectively, which are used in the \code{fit_sdl} function.}
}
\value{
The \code{sdlrm} function returns an object of class "sdlrm", which consists of a list with the following components.
\describe{
  \item{coefficients}{ a list containing the elements "mean" and
        "dispersion" that consist of the estimates of the
         coefficients associated with the mean and the dispersion
         parameter, respectively,}
  \item{link}{the link function used for the dispersion parameter model,}
  \item{vcov}{asymptotic covariance matrix of the maximum likelihood estimator of the model parameters vector,}
  \item{logLik}{log0likelihood of the ditted model,}
  \item{AIC}{model Akaike information criteria,}
  \item{BIC}{model bayesian information criteria,}
  \item{n.obs, p, k}{Sample size, number of coefficients in the mean model, and number of coefficients in the dispersion parameter model, respectively,}
  \item{feasible}{Logical. If \code{TRUE}, the estimates obtained belong to the parametric space,}
  \item{pearson.residuals}{a vector with the Pearson residuals,}
  \item{residuals}{a vector with the randomized quantile residuals,}
  \item{fitted.values}{a vector with the fitted means,}
  \item{phi.hat}{a vector with the fitted dispersion parameters,}
  \item{eta}{a vector with the fitted linear predictor for the dispersion parameters,}
  \item{response}{the vector of the response,}
  \item{Z, X}{model matrices associated with the mean and the dispersion parameter, respectively,}
  \item{call}{the function call,}
  \item{formula}{the formula used to specify the model in \code{sdlrm}.}
 }
}
\description{
Fit of the Sdl regression model via maximum
likelihood for a new parameterization of this distribution
that is indexed by the mean and a dispersion parameter.
}
\examples{

# Data set: pss (for description run ?pss)
barplot(table(pss$Diff), xlab = "PSS index difference", ylab = "Frequency")
boxplot(pss$Diff ~ pss$Group, xlab = "Group", ylab = "PSS index difference")

# Fit with a model only for the mean
fit_m <- sdlrm(Diff ~ Group, data = pss)
summary(fit_m)

# Fit a double model (mean and dispersion)
fit_md <- sdlrm(Diff ~ Group | Group, data = pss)
summary(fit_md)

# Fit a double model with a test for constant dispersion
fit <- sdlrm(Diff ~ Group | Group, data = pss, disp_test = TRUE)
summary(fit)
}
\references{
Medeiros, R. M. R., Bourguignon, M. (2020). Mean and
    dispersion regression model for paired count data with
    application to quality of life in prisons.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
