% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdl_estimation.R
\name{sdlrm_estimation}
\alias{sdlrm_estimation}
\alias{mle_sdl}
\title{Maximum likelihood estimates for the Sdl regression coefficients}
\usage{
mle_sdl(y, X, Z, link = "log", control = sdl_control(...),
  eq_constraint = NULL, eq_constraint_jac = NULL, ...)
}
\arguments{
\item{y}{a numeric integer-valued vector of the response variable.}

\item{X, Z}{model matrices associated with the mean and the dispersion parameter, respectively.}

\item{link}{character specification of the link function for the dispersion parameter. The links \code{"log"} (default) \code{"sqrt"} and \code{"identity"} can be used.}

\item{control}{a list of control arguments specified via \code{sdl_control}.}

\item{eq_constraint}{a function to evaluate equality constraints under the null hypothesis, in hypothesis testing contexts.}

\item{eq_constraint_jac}{a function to evaluate the jacobian of the equality constraints, , which were passed by the \code{eq_constraint} argument.}

\item{...}{arguments passed to \code{sdl_control}.}
}
\value{
A list with the values of the maximum likelihood estimates (\code{est});
the value of the log likelihood at estimates (\code{logLik}); and a logical value,
where \code{TRUE} indicates that the point obtained is feasible (\code{feasible}).
}
\description{
Maximum likelihood estimates for the coefficients of a fit
of the Sdl regression model.
}
\references{
Medeiros, R. M. R., Bourguignon, M. (2020). Mean and
    dispersion regression model for paired count data with
    application to quality of life in prisons.
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
